requirements.txt
anyio==3.6.1
click==8.1.3
fastapi==0.85.0
h11==0.14.0
idna==3.4
pydantic==1.10.2
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.4.0
uvicorn==0.18.3
psycopg2-binary==2.9.4
redis==4.3.4
databases==0.6.1
asyncpg==0.26.0
asyncio==3.4.3
environs==9.5.0
aioredis==2.0.1
typing==3.7.4.3
email-validator==1.3.0
sqlalchemy==1.4.41
alembic==1.8.1
fastapi-pagination==0.10.0
passlib==1.7.4
pyjwt==2.6.0
cryptography==38.0.1
python-jose==3.3.0
